docs/legacy/original_source/requirements.txt
# ── 核心依賴 ─────────────────────────────────────────────────────
anthropic>=0.40
vectorbt>=0.26
pandas>=2.0
pandas-ta>=0.3.14b
duckdb>=0.10
numpy>=1.26

# ── 資料庫 ───────────────────────────────────────────────────────
sqlalchemy>=2.0          # SQLite ORM（strategies / backtest_results / alerts）

# ── Web API ──────────────────────────────────────────────────────
fastapi>=0.110
uvicorn>=0.29
pydantic>=2.0            # FastAPI schema 驗證

# ── 報表 ─────────────────────────────────────────────────────────
plotly>=5.20
jinja2>=3.1
weasyprint>=62.0         # PDF 報表（選用）

# ── 工具 ─────────────────────────────────────────────────────────
python-dotenv>=1.0
pyyaml>=6.0
httpx>=0.27              # async HTTP client

# ── 測試 ─────────────────────────────────────────────────────────
pytest>=8.0

# ── Phase 2（預先列入，Phase 2 實作時啟用）───────────────────────
# celery>=5.3            # async task queue
# chromadb>=0.4          # RAG 向量資料庫
# sentence-transformers>=2.7  # 文字嵌入

# ── MultiCharts COM bridge（Windows only，Phase 2）───────────────
# pywin32>=306           # COM 介面（pip install pywin32）
# pywinauto>=0.6         # GUI 自動化（pip install pywinauto）
